项目文件小调整
.idea不应该被添加进git,
gradle和android sdk版本改为最高，
包名改成cn.atd3开头，
