+++ .gitignore
@@ -1,52 +1,24 @@
-<<<<<<< HEAD
-*.iml
-.gradle
-/local.properties
-/.idea/workspace.xml
-/.idea/libraries
+# General
 .DS_Store
+*~
+
+# Android Studio
+/.idea/
+/.gradle
+/local.properties
 /build
 /captures
-.externalNativeBuild
-=======
-# Built application files
+*.iml
+*.iws
+*.ipr
+*.swp
+
+# built application files
 *.apk
 *.ap_
 
-# Files for the ART/Dalvik VM
+# files for the dex VM
 *.dex
 
 # Java class files
 *.class
-
-# Generated files
-bin/
-gen/
-out/
-
-# Gradle files
-.gradle/
-build/
-
-# Local configuration file (sdk path, etc)
-local.properties
-
-# Proguard folder generated by Eclipse
-proguard/
-
-# Log Files
-*.log
-
-# Android Studio Navigation editor temp files
-.navigation/
-
-# Android Studio captures folder
-captures/
-
-# Intellij
-*.iml
-.idea/workspace.xml
-
-# Keystore files
-*.jks
->>>>>>> origin/master
