+++ app/src/main/java/cn/atd3/ygl/codemuseum/activity/useractivity/SettingActivity.java
@@ -1,0 +1,18 @@
+package cn.atd3.ygl.codemuseum.activity.useractivity;
+
+import android.os.Bundle;
+import android.support.v7.app.AppCompatActivity;
+
+import cn.atd3.ygl.codemuseum.R;
+
+/**
+ * Created by YGL on 2017/2/20.
+ */
+
+public class SettingActivity extends AppCompatActivity{
+    @Override
+    protected void onCreate(Bundle sls){
+        super.onCreate(sls);
+        setContentView(R.layout.settingactivity_layout);
+    }
+}
