+++ app/src/main/java/cn/atd3/ygl/codemuseum/util/HttpCallbackListener.java
@@ -1,0 +1,10 @@
+package cn.atd3.ygl.codemuseum.util;
+
+/**
+ * Created by YGL on 2017/2/13.
+ */
+
+public interface HttpCallbackListener {
+    void onFinish(String response);
+    void onError(Exception e);
+}
