+++ app/src/main/java/cn/atd3/ygl/codemuseum/util/HttpPictureCallbackListener.java
@@ -1,0 +1,12 @@
+package cn.atd3.ygl.codemuseum.util;
+
+import android.graphics.Bitmap;
+
+/**
+ * Created by YGL on 2017/2/25.
+ */
+
+public interface HttpPictureCallbackListener {
+    void onFinish(Bitmap bitmap);
+    void onError(Exception e);
+}
