--- app/src/main/java/com/example/ygl/codemuseum/activity/MainActivity.java
@@ -1,86 +1,0 @@
-package com.example.ygl.codemuseum.activity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.support.design.widget.NavigationView;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.MenuItem;
-import android.view.View;
-import android.widget.Button;
-
-import com.example.ygl.codemuseum.R;
-import com.example.ygl.codemuseum.activity.useractivity.MessageActivity;
-import com.example.ygl.codemuseum.activity.useractivity.SettingActivity;
-import com.example.ygl.codemuseum.activity.useractivity.SigninActivity;
-
-/**
- * Created by YGL on 2017/2/22.
- */
-
-public class MainActivity extends AppCompatActivity
-        implements NavigationView.OnNavigationItemSelectedListener {
-    @Override
-    protected void onCreate(Bundle sls){
-        super.onCreate(sls);
-        setContentView(R.layout.mainactivity_layout);
-
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawerlayout);
-        ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(
-                this, drawer, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
-        drawer.setDrawerListener(toggle);
-        toggle.syncState();
-
-        NavigationView navigationView = (NavigationView) findViewById(R.id.nav_view);
-        navigationView.setNavigationItemSelectedListener(this);
-
-        View view=navigationView.getHeaderView(0);
-        Button login_or_reg=(Button)view.findViewById(R.id.login_or_register);
-        login_or_reg.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent=new Intent(MainActivity.this,SigninActivity.class);
-                startActivity(intent);
-            }
-        });
-    }
-
-    @SuppressWarnings("StatementWithEmptyBody")
-    @Override
-    public boolean onNavigationItemSelected(MenuItem item) {
-        // Handle navigation view item clicks here.
-        int id = item.getItemId();
-        switch (id){
-            case R.id.messagemenu:
-                Intent messageActivityIntent=new Intent(MainActivity.this,MessageActivity.class);
-                startActivity(messageActivityIntent);
-                break;
-            case R.id.settingmenu:
-                Intent settingActivityIntent=new Intent(MainActivity.this,SettingActivity.class);
-                startActivity(settingActivityIntent);
-                break;
-            default:
-                break;
-        }
-
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawerlayout);
-        drawer.closeDrawer(GravityCompat.START);
-        return true;
-    }
-
-    @Override
-    public void onBackPressed() {
-        DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawerlayout);
-        if (drawer.isDrawerOpen(GravityCompat.START)) {
-            drawer.closeDrawer(GravityCompat.START);
-        } else {
-            super.onBackPressed();
-        }
-    }
-}
