--- app/src/main/java/com/example/ygl/codemuseum/activity/useractivity/SettingActivity.java
@@ -1,18 +1,0 @@
-package com.example.ygl.codemuseum.activity.useractivity;
-
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-
-import com.example.ygl.codemuseum.R;
-
-/**
- * Created by YGL on 2017/2/20.
- */
-
-public class SettingActivity extends AppCompatActivity{
-    @Override
-    protected void onCreate(Bundle sls){
-        super.onCreate(sls);
-        setContentView(R.layout.settingactivity_layout);
-    }
-}
