+++ app/src/main/res/layout/app_bar_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.example.ygl.codemuseum.activity.MainActivity">
+    tools:context="cn.atd3.ygl.codemuseum.activity.MainActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
